refactor aggregate functions to use new format
