+++ docs/src/piccolo/query_clauses/group_by.rst
@@ -19,7 +19,7 @@ In the following query, we get a count of the number of bands per manager:
 
 .. code-block:: python
 
-    >>> from piccolo.query.methods.select import Count
+    >>> from piccolo.query.functions.aggregate import Count
 
     >>> await Band.select(
     ...     Band.manager.name.as_alias('manager_name'),
