+++ piccolo/query/__init__.py
@@ -1,9 +1,9 @@
 from piccolo.columns.combination import WhereRaw
 
 from .base import Query
+from .functions.aggregate import Avg, Max, Min, Sum
 from .methods import (
     Alter,
-    Avg,
     Count,
     Create,
     CreateIndex,
@@ -11,12 +11,9 @@ from .methods import (
     DropIndex,
     Exists,
     Insert,
-    Max,
-    Min,
     Objects,
     Raw,
     Select,
-    Sum,
     TableExists,
     Update,
 )
