+++ piccolo/query/functions/__init__.py
@@ -1,10 +1,16 @@
+from .aggregate import Avg, Count, Max, Min, Sum
 from .string import Length, Lower, Ltrim, Reverse, Rtrim, Upper
 
 __all__ = (
+    "Avg",
+    "Count",
     "Length",
     "Lower",
     "Ltrim",
+    "Max",
+    "Min",
     "Reverse",
     "Rtrim",
+    "Sum",
     "Upper",
 )
