+++ piccolo/query/methods/select.py
@@ -1,6 +1,5 @@
 from __future__ import annotations
 
-import decimal
 import itertools
 import typing as t
 from collections import OrderedDict
@@ -36,13 +35,8 @@ if t.TYPE_CHECKING:  # pragma: no cover
     from piccolo.custom_types import Combinable
     from piccolo.table import Table  # noqa
 
-
-def is_numeric_column(column: Column) -> bool:
-    return column.value_type in (int, decimal.Decimal, float)
-
-
-def is_string_column(column: Column) -> bool:
-    return column.value_type is str
+# Here to avoid breaking changes - will be removed in the future.
+from piccolo.query.functions.aggregate import Count  # noqa: F401
 
 
 class SelectRaw(Selectable):
@@ -65,222 +59,6 @@ class SelectRaw(Selectable):
         self, engine_type: str, with_alias: bool = True
     ) -> QueryString:
         return self.querystring
-
-
-class Avg(Selectable):
-    """
-    ``AVG()`` SQL function. Column type must be numeric to run the query.
-
-    .. code-block:: python
-
-        await Band.select(Avg(Band.popularity)).run()
-
-        # We can use an alias. These two are equivalent:
-
-        await Band.select(
-            Avg(Band.popularity, alias="popularity_avg")
-        ).run()
-
-        await Band.select(
-            Avg(Band.popularity).as_alias("popularity_avg")
-        ).run()
-
-    """
-
-    def __init__(self, column: Column, alias: str = "avg"):
-        if is_numeric_column(column):
-            self.column = column
-        else:
-            raise ValueError("Column type must be numeric to run the query.")
-        self._alias = alias
-
-    def get_select_string(
-        self, engine_type: str, with_alias: bool = True
-    ) -> QueryString:
-        column_name = self.column._meta.get_full_name(with_alias=False)
-        return QueryString(f'AVG({column_name}) AS "{self._alias}"')
-
-
-class Count(Selectable):
-    """
-    Used in ``Select`` queries, usually in conjunction with the ``group_by``
-    clause::
-
-        >>> await Band.select(
-        ...     Band.manager.name.as_alias('manager_name'),
-        ...     Count(alias='band_count')
-        ... ).group_by(Band.manager)
-        [{'manager_name': 'Guido', 'count': 1}, ...]
-
-    It can also be used without the ``group_by`` clause (though you may prefer
-    to the :meth:`Table.count <piccolo.table.Table.count>` method instead, as
-    it's more convenient)::
-
-        >>> await Band.select(Count())
-        [{'count': 3}]
-
-    """
-
-    def __init__(
-        self,
-        column: t.Optional[Column] = None,
-        distinct: t.Optional[t.Sequence[Column]] = None,
-        alias: str = "count",
-    ):
-        """
-        :param column:
-            If specified, the count is for non-null values in that column.
-        :param distinct:
-            If specified, the count is for distinct values in those columns.
-        :param alias:
-            The name of the value in the response::
-
-                # These two are equivalent:
-
-                await Band.select(
-                    Band.name, Count(alias="total")
-                ).group_by(Band.name)
-
-                await Band.select(
-                    Band.name,
-                    Count().as_alias("total")
-                ).group_by(Band.name)
-
-        """
-        if distinct and column:
-            raise ValueError("Only specify `column` or `distinct`")
-
-        self.column = column
-        self.distinct = distinct
-        self._alias = alias
-
-    def get_select_string(
-        self, engine_type: str, with_alias: bool = True
-    ) -> QueryString:
-        expression: str
-
-        if self.distinct:
-            if engine_type == "sqlite":
-                # SQLite doesn't allow us to specify multiple columns, so
-                # instead we concatenate the values.
-                column_names = " || ".join(
-                    i._meta.get_full_name(with_alias=False)
-                    for i in self.distinct
-                )
-            else:
-                column_names = ", ".join(
-                    i._meta.get_full_name(with_alias=False)
-                    for i in self.distinct
-                )
-
-            expression = f"DISTINCT ({column_names})"
-        else:
-            if self.column:
-                expression = self.column._meta.get_full_name(with_alias=False)
-            else:
-                expression = "*"
-
-        return QueryString(f'COUNT({expression}) AS "{self._alias}"')
-
-
-class Max(Selectable):
-    """
-    ``MAX()`` SQL function.
-
-    .. code-block:: python
-
-        await Band.select(
-            Max(Band.popularity)
-        ).run()
-
-        # We can use an alias. These two are equivalent:
-
-        await Band.select(
-            Max(Band.popularity, alias="popularity_max")
-        ).run()
-
-        await Band.select(
-            Max(Band.popularity).as_alias("popularity_max")
-        ).run()
-
-    """
-
-    def __init__(self, column: Column, alias: str = "max"):
-        self.column = column
-        self._alias = alias
-
-    def get_select_string(
-        self, engine_type: str, with_alias: bool = True
-    ) -> QueryString:
-        column_name = self.column._meta.get_full_name(with_alias=False)
-        return QueryString(f'MAX({column_name}) AS "{self._alias}"')
-
-
-class Min(Selectable):
-    """
-    ``MIN()`` SQL function.
-
-    .. code-block:: python
-
-        await Band.select(Min(Band.popularity)).run()
-
-        # We can use an alias. These two are equivalent:
-
-        await Band.select(
-            Min(Band.popularity, alias="popularity_min")
-        ).run()
-
-        await Band.select(
-            Min(Band.popularity).as_alias("popularity_min")
-        ).run()
-
-    """
-
-    def __init__(self, column: Column, alias: str = "min"):
-        self.column = column
-        self._alias = alias
-
-    def get_select_string(
-        self, engine_type: str, with_alias: bool = True
-    ) -> QueryString:
-        column_name = self.column._meta.get_full_name(with_alias=False)
-        return QueryString(f'MIN({column_name}) AS "{self._alias}"')
-
-
-class Sum(Selectable):
-    """
-    ``SUM()`` SQL function. Column type must be numeric to run the query.
-
-    .. code-block:: python
-
-        await Band.select(
-            Sum(Band.popularity)
-        ).run()
-
-        # We can use an alias. These two are equivalent:
-
-        await Band.select(
-            Sum(Band.popularity, alias="popularity_sum")
-        ).run()
-
-        await Band.select(
-            Sum(Band.popularity).as_alias("popularity_sum")
-        ).run()
-
-    """
-
-    def __init__(self, column: Column, alias: str = "sum"):
-        if is_numeric_column(column):
-            self.column = column
-        else:
-            raise ValueError("Column type must be numeric to run the query.")
-        self._alias = alias
-
-    def get_select_string(
-        self, engine_type: str, with_alias: bool = True
-    ) -> QueryString:
-        column_name = self.column._meta.get_full_name(with_alias=False)
-        return QueryString(f'SUM({column_name}) AS "{self._alias}"')
 
 
 OptionalDict = t.Optional[t.Dict[str, t.Any]]
